requirements.txt
APScheduler==3.10.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
django-environ==0.10.0
idna==3.4
jdatetime==4.1.1
mysql-connector==2.2.9
python-telegram-bot==20.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
tornado==6.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
